synapses/telegram-bot/pyproject.toml
[project]
name = "telegram-bot"
version = "0.1.0"
description = "Telegram Bot for Aura Platform"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.17.0",
    "grpcio>=1.70.0",
    "protobuf>=6.33.5",
    "pydantic-settings>=2.7.1",
    "structlog>=25.1.0",
    "opentelemetry-api>=1.24.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-exporter-otlp>=1.24.0",
    "nats-py>=2.9.0",
    "aura-core",
    "core",
    "betterproto>=2.0.0b7",
]

[tool.uv.sources]
aura-core = { workspace = true }
core = { workspace = true }

[dependency-groups]
dev = [
    "grpcio-tools>=1.70.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
]

[tool.uv]
managed = true
